pyproject.toml
[tool.poetry]
name = "doverchat"
version = "0.1.0"
description = ""
authors = ["Logan Yang <logancyang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-socketio = "^4.3.1"
eventlet = "^0.25.2"
python-dotenv = "^0.14.0"
gunicorn = "^20.0.4"
flask-login = "^0.5.0"
sqlalchemy = "^1.3.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
